backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.10.3
python-dotenv==1.0.0
certifi==2024.2.2
motor==3.1.1
pymongo[srv]==4.3.3
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
